Added some more comments

+++ skinlib/src/main/java/net/zatrit/skins/lib/Config.java
@@ -16,7 +16,7 @@ import java.util.concurrent.Executors;
 public final class Config {
     private final Gson gson = new GsonBuilder().create();
     private @Setter @Nullable CacheProvider cacheProvider;
-    private @Setter int loaderTimeout;
+    private @Setter float loaderTimeout;
     private @Builder.Default Executor executor = Executors.newFixedThreadPool(
             Runtime.getRuntime().availableProcessors());
 }
